refactor: move SomeNumberCell to its own file
